feat: Extract initNode to own folder
